image-builder: Use epoch+1 as epoch is messy
KDE things may fail in almost undebuggable ways.

With "firehose log level" I saw this message for files that were
present:

https://invent.kde.org/frameworks/kservice/-/blob/9eb46f9b20eb5934b81cab9871ce1a0c0604bbe6/src/sycoca/kbuildsycoca.cpp#L636

This, in turn, is because "hash" is computed as a sum of all modified
dates

https://invent.kde.org/frameworks/kservice/-/blob/9eb46f9b20eb5934b81cab9871ce1a0c0604bbe6/src/sycoca/kbuildsycoca.cpp#L610

(Weird hashing!)

Which, in turn, meant that it was `0` here.

Finally, `0` ends up magically breaking everything in a plasma mobile
system. Since apparently desktop files will not even be attempted to be
read or used if they are not cached by ksycoca.

The task panel broke in inscrutable ways, and there were no applications
found to run.