STATE: Abandoned; hardware does not allow unlocking
In theory this branch, as it stands, may produce a working boot image.

At the very least, it compiles in full.

Though, it is impossible to say since ZTE devices, I learned after the
fact, generally do not allow bootloader unlocking to happen. That is,
**even** if the OEM unlocking toggle is present in developer settings!

The ZTE aboot implementation (fastboot) has *all commands* disabled, and
not only for this phone, from research. This means that `oem unlock` is
not available, and some Ghidra spelunking confirmed this. Even though
some messages and stubs exist for unlocking, the actual place where
rubber meets the road, the *commands handling* does not exist.

As such, this is a theoretical port. This could have been an amazing
alternative OS for this phone, but alas, we'll never know!