[HACK] pine64-pinephonepro: Add workaround to show the logo
This serves more to prove to me that it actually should be working. What
needs to actually be done is figure out *what* is preventing the logo
from showing.

Note that this is not a Pinephone Pro issue, nor is it an RK3399 issue,
nor is it an ARM issue. This workaround had been authored for an x86_64
system initially.