mobile-nixos.kernel-append-dtbs: Forward override and version
The override method shown here is a bit odd, but allows forwarding the
arguments to the current call to an "overriden" kernel.

In other words, we get the semantics of `override` on a kernel, which is
kinda hard to get correct, for free. Though we re-apply the wrap
afterward.

This does not incurr an additional build to the kernel, as long as the
`override` call wouldn't otherwise.