pine64-pinenote: Add orientation patch to the kernel
Before upstreaming, it will be important to entirely understand why the
default is to reflect the panel.

Anyway, what this does is to give the correct hints such that DRM-using
apps can render in the correct orientation.

This is needed for the Mobile NixOS stage-1 user interface, or else it's
awkwardly set in the wrong orientation.