pine64-pinenote: Make cover case act as lid switch
This provides automatic support of proper resume/suspend sequence by
software like logind.

Without this, having the cover case closed, and connecting to power
would wake the device and keep it woken.

With software that knows about lid switch, it can be configured to
correctly go back to suspend when closed.