kernel/builder: Accept logo from pkgs
There is no `override` that will work given the number of `callPackage`
calls between this and the actual derivation.

I guess `overrideAttrs` could be used instead, if we used an attribute
of the derivation instead. I'm not sure if it's better since it would
entail using `apply` to `overrideAttrs` the final `kernel.package`
value.