modules/initrd-kernel: Resize the logo according to the display size
By using a square aspect ratio logo as the source material, we can
"fill" the area, which will not go outside of the bounds, which implies
that the image is scaled to the smaller dimension.

This image is then trimmed, meaning that it is cut to remove any
redundant colour at the outside edge.