modules/boot: Add options to better control console parameters
It is generally awkward to deal with the console kernel parameters, and
just as awkward to deal with lists in the modules system.

Internally to Mobile NixOS, we'll use these values that should always be
prefixed to any other values. This means that those consoles will likely
to be less preferred than any the end-users sets (e.g. `console=tty1` in
their own config).

The option for `defaultConsole` is not really meant to be
user-controlled, but there is actually no issue if they use it, as long
as their configuration does not set another console later in the
cmdline.