support/nix: Add pure-eval.sh
This script helps work around the issues that pure evaluation is really
inconvenient outside of Flakes. Especially for development purpose.

This is an actual `pure-eval`, though, which means it should be working
the same when coupled with a less synthetic way to refer to the Mobile
NixOS repository... except if paths end-up not being made available, but
that's kind of out of scope for this tool.

This tool is planned to be added to CI at some point to catch bad evals.