support/nix: Add forbid-builtins
This is a somewhat more convenient, but less strong, way to check for
impurities during an evaluation.

The main issue here is it's faking it by using `scopedImport`. It does
not respect other properties from a pure eval.

With that said, it can be used to forbid additional stuff being used
within Mobile NixOS, so it's worthwhile to also include as a tool.

And it's why it's used to check there's no other unexpected inputs.